// PRN232_FE/src/types/property.ts
export type Property = {
  id: number
  name: string
  address: string
  distance: string
  price: string
  rooms: number
  image: string
  x: string
  y: string
  amenities: string[]
  description: string
}

export type NearbyPlace = {
  name: string
  category: string
  distance: string
  travel: string
  x: string
  y: string
}